requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
youtube-dl==2021.12.17
yt-dlp==2023.12.30
python-dotenv==1.0.0
spotipy==2.23.0
requests==2.31.0
